move python into genny subdir to make loggers clearer
