+++ src/lamplib/genny/tasks/auto_tasks.py
@@ -14,7 +14,7 @@ from shrub.command import CommandDefinition
 from shrub.config import Configuration
 from shrub.variant import TaskSpec
 
-from cmd_runner import run_command
+from genny.cmd_runner import run_command
 
 SLOG = structlog.get_logger(__name__)
