+++ src/lamplib/genny/tasks/generate-uuid-tag.sh
