+++ src/lamplib/genny/tasks/genny_runner.py
@@ -3,8 +3,8 @@ from typing import List
 import structlog
 import os
 
-from cmd_runner import run_command
-from curator import poplar_grpc
+from genny.cmd_runner import run_command
+from genny.curator import poplar_grpc
 
 SLOG = structlog.get_logger(__name__)
